refactor(forecaster): extract month_utils, aggregators, history_loaders from cli.py
Pure code motion, zero behavior change (6,952 -> 5,742 lines in cli.py):
- forecaster/month_utils.py (222 LOC): month/window arithmetic
  (_anchor_month_for_question, _month_index_for_label, _expected_months,
  key parsers).
- forecaster/aggregators.py (186 LOC): calibration-weight loading + the
  weights cache/lock (object identity preserved — cli re-imports the
  same dict/lock so cli._CALIB_WEIGHTS_CACHE.clear() still works).
- forecaster/history_loaders.py (908 LOC): IFRC/IDMC/ACLED/GDACS history
  and base-rate builders + _format_base_rate_for_prompt.

cli.py re-imports every moved name (backward-compat re-exports), so all
existing 'from forecaster.cli import X' imports and cli.X references
keep working. The writer cluster (_write_spd_outputs,
_write_binary_outputs, _write_spd_members_v2_to_db) and the
history-summary/BayesMC callers deliberately STAYED in cli.py: tests
monkeypatch cli.connect / cli.duckdb / cli._load_calibration_weights_db /
cli.aggregate_spd_v2_bayesmc, and those functions must resolve the
patched names through cli's namespace (moving them would silently break
the patch seams). CLAUDE.md documents the new layout and the
do-not-move seam list.

Test suites identical to baseline before/after (169 passed, 6
pre-existing failures); zero test files modified.
