+++ CLAUDE.md
@@ -92,7 +92,7 @@ Pythia/
 - `horizon_scanner/seasonal_tc/__init__.py` — Country-to-basin mapping + cached TC context reader
 - `horizon_scanner/hdx_signals.py` — HDX Signals connector (OCHA automated crisis monitoring, indicator-to-hazard mapping)
 - `horizon_scanner/conflict_forecasts.py` — Unified conflict forecast loader (VIEWS + conflictforecast.org + ACLED CAST)
-- `forecaster/cli.py` — Forecaster main runner (~7300 LOC). Supports both SPD (per-metric bucket count × 6-month; 6 buckets for PA/PHASE3PLUS_IN_NEED, 7 for FATALITIES) and binary (per-month probability) forecast pipelines. Binary questions (metric=EVENT_OCCURRENCE) are detected and routed to `_run_binary_forecast_for_question()`, which uses `build_binary_event_prompt()` for prompting and `parse_binary_response()` for output parsing. Binary storage convention: bucket_1 = P(yes), bucket_2 = P(no) = 1-P(yes), remaining buckets = 0 (EVENT_OCCURRENCE is not in BUCKET_SPECS; the binary writer is independent of the SPD bucket schemes). History builders: `_build_history_summary` dispatches to `_load_fewsnet_phase3_history` for DR/PHASE3PLUS_IN_NEED (null-aware, coverage-tracked), `_build_natural_hazard_seasonal_profile` for natural hazard PA, `_build_conflict_base_rate` for ACE, and `_build_gdacs_event_history` for FL/DR/TC event occurrence (binary 1/0 + alertlevel from facts_resolved). `_format_base_rate_for_prompt` handles `fewsnet_phase3` type with null-aware formatting showing "null" for FEWS NET analysis cycle gaps. GDACS event history is loaded in `_load_structured_data()` for FL/DR/TC hazards and injected into both SPD prompts (via `_format_gdacs_event_history_for_prompt` in prompts.py) and binary prompts (via `gdacs_event_history` parameter on `build_binary_event_prompt`).
+- `forecaster/cli.py` — Forecaster main runner (~5700 LOC after the July 2026 decomposition). **Decomposed (pure code motion, zero behavior change)** into three sibling modules, with cli.py re-importing every moved name as a backward-compat re-export (`forecaster.cli.X` and `from forecaster.cli import X` still work; the mutable `_CALIB_WEIGHTS_CACHE`/`_CALIB_WEIGHTS_LOCK` keep object identity): `forecaster/month_utils.py` (month/window arithmetic: `_anchor_month_for_question`, `_month_index_for_label`, `_expected_months`, `_parse_month_key`, `_add_months`, `_sanitize_month_series`, etc.), `forecaster/history_loaders.py` (IFRC/IDMC/ACLED PA history loaders, `_load_idmc_conflict_flow_history_summary`, `_build_gdacs_event_history`, `_format_base_rate_for_prompt`, `_load_country_names`, `NATURAL_HAZARD_CODES`/`IDMC_HZ_MAP`), and `forecaster/aggregators.py` (calibration-weight loading/caching + `_pythia_db_url_from_config`/`_pythia_db_path_from_config`). **Intentionally NOT moved** because tests monkeypatch cli-module attributes that the callees resolve at call time (moving them would silently break the patches): the DB writers `_write_spd_outputs`/`_write_binary_outputs`/`_write_spd_members_v2_to_db` (tests patch `forecaster.cli.connect`; writers also read the module-global `_IS_TEST`), `_build_history_summary` + `_build_natural_hazard_seasonal_profile` + `_build_conflict_base_rate` (patch `cli.connect`), `_load_fewsnet_phase3_history` (patch `cli.duckdb` + `cli._pythia_db_path_from_config`), `_resolve_member_weights` (patch `cli._load_calibration_weights_db`), and `_build_bayesmc_spd_obj` (patch `cli.aggregate_spd_v2_bayesmc`) — do not move these without reworking those test seams. Supports both SPD (per-metric bucket count × 6-month; 6 buckets for PA/PHASE3PLUS_IN_NEED, 7 for FATALITIES) and binary (per-month probability) forecast pipelines. Binary questions (metric=EVENT_OCCURRENCE) are detected and routed to `_run_binary_forecast_for_question()`, which uses `build_binary_event_prompt()` for prompting and `parse_binary_response()` for output parsing. Binary storage convention: bucket_1 = P(yes), bucket_2 = P(no) = 1-P(yes), remaining buckets = 0 (EVENT_OCCURRENCE is not in BUCKET_SPECS; the binary writer is independent of the SPD bucket schemes). History builders: `_build_history_summary` dispatches to `_load_fewsnet_phase3_history` for DR/PHASE3PLUS_IN_NEED (null-aware, coverage-tracked), `_build_natural_hazard_seasonal_profile` for natural hazard PA, `_build_conflict_base_rate` for ACE, and `_build_gdacs_event_history` for FL/DR/TC event occurrence (binary 1/0 + alertlevel from facts_resolved). `_format_base_rate_for_prompt` handles `fewsnet_phase3` type with null-aware formatting showing "null" for FEWS NET analysis cycle gaps. GDACS event history is loaded in `_load_structured_data()` for FL/DR/TC hazards and injected into both SPD prompts (via `_format_gdacs_event_history_for_prompt` in prompts.py) and binary prompts (via `gdacs_event_history` parameter on `build_binary_event_prompt`).
 - `pythia/api/app.py` — FastAPI application (~3500 LOC). Phase 4 additions: `_get_risk_index_binary()` serves EVENT_OCCURRENCE probabilities (bucket_1 = P(event)), `/v1/diagnostics/resolution_rates` endpoint returns resolution rate by (hazard, metric). Risk index supports EVENT_OCCURRENCE (binary), PHASE3PLUS_IN_NEED (SPD), PA, and FATALITIES metrics. KPI scopes accept all four metric scopes. **Resolver accordion endpoints**: `/v1/resolver/source_inventory` returns per-source metadata (label, category, last_updated, row counts); `/v1/resolver/source_data` returns rows for a specific source with curated or all columns. Source registry (`_SOURCE_REGISTRY`) maps 22 source keys to tables, filters, curated columns. `facts_resolved` sources filtered by `publisher` column (IFRC, IDMC, ACLED, GDACS / JRC, FEWS NET). Last-updated dates use ingestion-time columns (`created_at`, `fetched_at`) not data-period columns. **`/v1/version`** live-computes staleness fields from the synced DB: `latest_forecast_month`/`latest_forecast_run_id` plus `latest_scores_at`, `latest_calibration_at`, and `latest_data_at` (= `MAX(created_at)` across `latest_hs_created_at`/`forecasts_ensemble`/`resolutions`/`scores`/`calibration_weights`/`calibration_advice`, excluding the manifest's publish-time `created_utc`). The dashboard homepage "Last updated" renders `latest_data_at` so the scoring/calibration loop counts as a data refresh even when it creates no new HS run; the HS-run date is shown separately as "Latest forecast scan".
 - `web/src/app/questions/[questionId]/BinaryPanel.tsx` — Binary event forecast display: per-month probability bars, resolution outcomes (event/no event), per-horizon Brier scores.
 - `pythia/tools/compute_resolutions.py` — Resolves forecasts against Resolver ground truth. Supports PA, FATALITIES, EVENT_OCCURRENCE, and PHASE3PLUS_IN_NEED metrics. Source-aware null handling: FATALITIES+ACE/ACO and EVENT_OCCURRENCE default to 0 when no data; PA and PHASE3PLUS_IN_NEED leave horizons unresolved (null) when no data exists. FEWS NET IPC resolution via `_try_fewsnet_ipc()` queries only `phase3plus_in_need` metric (not projections), guarded to DR hazard only.
