+++ forecaster/cli.py
@@ -70,11 +70,55 @@ from forecaster.self_search import (
     trim_sources,
 )
 
+# --- Backward-compat re-exports (pure code motion) --------------------------
+# These names were moved verbatim to forecaster/month_utils.py,
+# forecaster/aggregators.py and forecaster/history_loaders.py. They are
+# re-imported here so `forecaster.cli.X` and `from forecaster.cli import X`
+# keep working for every moved name (including the module-level mutable
+# cache/lock, whose object identity is preserved).
+from forecaster.month_utils import (  # noqa: F401 - backward-compat re-exports
+    _add_months,
+    _anchor_month_for_question,
+    _coerce_date,
+    _expected_months,
+    _first_target_month,
+    _is_calendar_month_key,
+    _month_index_for_label,
+    _parse_month_key,
+    _parse_month_offset_key,
+    _sanitize_month_series,
+)
+from forecaster.aggregators import (  # noqa: F401 - backward-compat re-exports
+    _CALIB_WEIGHTS_CACHE,
+    _CALIB_WEIGHTS_LOCK,
+    _PYTHIA_CFG_LOAD,
+    _calibration_weights_enabled,
+    _load_calibration_weights_db,
+    _load_calibration_weights_file,
+    _member_weight_key,
+    _pythia_db_path_from_config,
+    _pythia_db_url_from_config,
+    _reset_calibration_weights_cache,
+)
+from forecaster.history_loaders import (  # noqa: F401 - backward-compat re-exports
+    COUNTRIES_CSV,
+    IDMC_HZ_MAP,
+    NATURAL_HAZARD_CODES,
+    _HAZARD_DISPLAY_NAMES,
+    _build_gdacs_event_history,
+    _format_base_rate_for_prompt,
+    _load_acled_fatalities_history,
+    _load_country_names,
+    _load_idmc_conflict_flow_history_summary,
+    _load_idmc_pa_history,
+    _load_ifrc_pa_history,
+    _load_pa_history_block,
+)
+
 LOG = logging.getLogger(__name__)
 
 MAX_RESEARCH_WORKERS = int(os.getenv("FORECASTER_RESEARCH_MAX_WORKERS", "6"))
 MAX_SPD_WORKERS = int(os.getenv("FORECASTER_SPD_MAX_WORKERS", "6"))
-COUNTRIES_CSV = Path(__file__).resolve().parents[1] / "resolver" / "data" / "countries.csv"
 
 
 _DEFAULT_ENSEMBLE_LOGGED = False
@@ -209,10 +253,6 @@ def _dedupe_sources_by_url(sources: list[Any]) -> list[dict[str, Any]]:
     return deduped
 
 
-_PYTHIA_CFG_LOAD = None
-if importlib.util.find_spec("pythia.config") is not None:
-    _PYTHIA_CFG_LOAD = getattr(importlib.import_module("pythia.config"), "load", None)
-
 try:
     from pythia.llm_profiles import get_current_models as _get_llm_profile_models
 except Exception:
@@ -266,194 +306,10 @@ HZ_QUERY_MAP = {
     "DI": "DISPLACEMENT",
 }
 
-NATURAL_HAZARD_CODES = {"FL", "DR", "TC", "HW"}
-
-IDMC_HZ_MAP = {"ACE", "DI"}
-
-
-def _load_idmc_conflict_flow_history_summary(
-    con,
-    iso3: str,
-    hazard_code: str,
-) -> Dict[str, Any]:
-    """
-    Best-effort IDMC/DTM conflict displacement history for use in _build_history_summary.
-
-    Reads per-month flows from facts_deltas for conflict hazards:
-      - metric ∈ {new_displacements, idp_displacement_new_dtm, idp_displacement_flow_idmc}
-      - series_semantics = 'new'
-      - iso3 and hazard_code filtered.
-
-    Returns a summary dict with:
-      - source: "IDMC"
-      - history_length_months: number of months with non-zero flow
-      - recent_mean: mean over the last up-to-6 months
-      - recent_max: max over the full window
-      - trend: crude "increasing" / "decreasing" / "flat" based on first vs last month
-      - last_6m_values: list of {"ym", "value"} for up to the last 6 months
-      - data_quality: "medium" (we’re using delta flows only)
-      - notes: short explanation of what we used.
-    """
-    iso3_up = (iso3 or "").upper().strip()
-    hz_up = (hazard_code or "").upper().strip()
-
-    if not iso3_up or not hz_up:
-        return {
-            "source": "IDMC",
-            "history_length_months": 0,
-            "recent_mean": None,
-            "recent_max": None,
-            "trend": "uncertain",
-            "last_6m_values": [],
-            "data_quality": "low",
-            "notes": "Missing iso3 or hazard_code for IDMC conflict PA history.",
-        }
-
-    try:
-        rows = con.execute(
-            """
-            SELECT
-                ym,
-                SUM(
-                    CASE
-                        WHEN lower(metric) = 'new_displacements'
-                             THEN COALESCE(value_new, 0)
-                        WHEN lower(metric) = 'idp_displacement_new_dtm'
-                             THEN COALESCE(value_new, 0)
-                        WHEN lower(metric) = 'idp_displacement_flow_idmc'
-                             THEN COALESCE(value_new, 0)
-                        ELSE 0
-                    END
-                ) AS flow_value
-            FROM facts_deltas
-            WHERE upper(iso3) = ?
-              AND COALESCE(NULLIF(upper(hazard_code), ''), 'ACE') IN (?, 'IDU')
-              AND lower(series_semantics) = 'new'
-              AND lower(metric) IN (
-                'new_displacements',
-                'idp_displacement_new_dtm',
-                'idp_displacement_flow_idmc'
-              )
-            GROUP BY ym
-            ORDER BY ym
-            """,
-            [iso3_up, hz_up],
-        ).fetchall()
-    except Exception as exc:  # noqa: BLE001
-        logging.warning(
-            "IDMC conflict PA history query failed for %s/%s: %s",
-            iso3_up,
-            hz_up,
-            exc,
-        )
-        return {
-            "source": "IDMC",
-            "history_length_months": 0,
-            "recent_mean": None,
-            "recent_max": None,
-            "trend": "uncertain",
-            "last_6m_values": [],
-            "data_quality": "low",
-            "notes": f"IDMC history unavailable due to query error: {type(exc).__name__}",
-        }
-
-    logging.debug("IDMC displacement query for %s/%s: %d rows", iso3_up, hz_up, len(rows))
-
-    if not rows:
-        return {
-            "source": "IDMC",
-            "history_length_months": 0,
-            "recent_mean": None,
-            "recent_max": None,
-            "trend": "uncertain",
-            "last_6m_values": [],
-            "data_quality": "low",
-            "notes": "No IDMC/DTM displacement flow history found in facts_deltas.",
-        }
-
-    month_to_value: Dict[str, float] = {}
-    for ym_val, flow_val in rows:
-        try:
-            v = float(flow_val or 0.0)
-        except Exception:
-            v = 0.0
-        if v == 0:
-            continue
-        month_to_value[str(ym_val)] = v
-
-    month_to_value, dropped_future, unparseable_keys = _sanitize_month_series(month_to_value)
-    if not month_to_value:
-        summary_empty: Dict[str, Any] = {
-            "source": "IDMC",
-            "history_length_months": 0,
-            "recent_mean": None,
-            "recent_max": None,
-            "trend": "uncertain",
-            "last_6m_values": [],
-            "data_quality": "low",
-            "notes": "IDMC displacement flows found but all were zero or filtered out.",
-        }
-        sanity_empty: Dict[str, Any] = {}
-        if dropped_future:
-            sanity_empty["dropped_future_months"] = dropped_future
-        if unparseable_keys:
-            sanity_empty["unparseable_month_keys"] = unparseable_keys
-        if sanity_empty:
-            summary_empty["_sanity"] = sanity_empty
-        return summary_empty
-
-    ordered_items = sorted(month_to_value.items())
-    history: list[Dict[str, Any]] = [{"ym": k, "value": v} for k, v in ordered_items]
-    values: list[float] = [float(v) for _k, v in ordered_items]
-
-    n = len(history)
-    last_6 = history[-6:]
-    recent_window = values[-6:]
-    recent_mean = sum(recent_window) / len(recent_window) if recent_window else None
-    recent_max = max(values) if values else None
-
-    trend = "flat"
-    if len(values) >= 2:
-        if values[-1] > values[0]:
-            trend = "increasing"
-        elif values[-1] < values[0]:
-            trend = "decreasing"
-
-    summary: Dict[str, Any] = {
-        "source": "IDMC",
-        "history_length_months": n,
-        "recent_mean": recent_mean,
-        "recent_max": recent_max,
-        "trend": trend,
-        "last_6m_values": last_6,
-        "data_quality": "medium",
-        "notes": (
-            "IDMC/DTM monthly displacement flows (facts_deltas) used as the conflict PA base rate."
-        ),
-    }
-    sanity: Dict[str, Any] = {}
-    if dropped_future:
-        sanity["dropped_future_months"] = dropped_future
-    if unparseable_keys:
-        sanity["unparseable_month_keys"] = unparseable_keys
-    if sanity:
-        summary["_sanity"] = sanity
-    return summary
-
 
 # ---------------------------------------------------------------------------
 # Hazard-specific base-rate builders (replace generic JSON dump in SPD v2)
 # ---------------------------------------------------------------------------
-
-_HAZARD_DISPLAY_NAMES: Dict[str, str] = {
-    "FL": "Flood",
-    "DR": "Drought",
-    "TC": "Tropical Cyclone",
-    "HW": "Heat Wave",
-    "ACE": "Armed Conflict",
-    "DI": "Displacement Inflow",
-}
-
 
 def _build_natural_hazard_seasonal_profile(
     iso3: str,
@@ -714,334 +570,6 @@ def _build_conflict_base_rate(
         }
     finally:
         con.close()
-
-
-def _build_gdacs_event_history(
-    iso3: str,
-    hazard_code: str,
-) -> Dict[str, Any] | None:
-    """Build GDACS event occurrence history for natural hazard base rate support.
-
-    Queries facts_resolved for event_occurrence metric (binary 1/0) and
-    alertlevel (Green/Orange/Red) for the given country + hazard.
-
-    Only applicable for FL, DR, TC hazards.
-
-    Returns None if no GDACS data exists for this country-hazard.
-    """
-    iso3_up = (iso3 or "").upper().strip()
-    hz_up = (hazard_code or "").upper().strip()
-
-    if hz_up not in ("FL", "DR", "TC"):
-        return None
-
-    con = connect(read_only=True)
-    try:
-        rows = con.execute(
-            """
-            SELECT ym, value, alertlevel
-            FROM facts_resolved
-            WHERE upper(iso3) = ?
-              AND upper(hazard_code) = ?
-              AND lower(metric) = 'event_occurrence'
-            ORDER BY ym
-            """,
-            [iso3_up, hz_up],
-        ).fetchall()
-    except Exception:
-        return None
-    finally:
-        con.close()
-
-    if not rows:
-        return None
-
-    # Parse into structured records
-    now_ym = datetime.now(timezone.utc).replace(tzinfo=None).strftime("%Y-%m")
-    events: list[dict] = []
-    for ym_raw, value, alertlevel in rows:
-        ym = str(ym_raw)[:7] if ym_raw else ""
-        if not ym or ym > now_ym:
-            continue
-        try:
-            v = float(value or 0)
-        except (TypeError, ValueError):
-            v = 0.0
-        events.append({
-            "ym": ym,
-            "occurred": v >= 1.0,
-            "alertlevel": str(alertlevel or "").strip() or None,
-        })
-
-    if not events:
-        return None
-
-    events_sorted = sorted(events, key=lambda e: e["ym"])
-
-    # Compute summary stats
-    total_months = len(events_sorted)
-    event_months = sum(1 for e in events_sorted if e["occurred"])
-    event_rate = event_months / total_months if total_months > 0 else 0.0
-
-    # Data range
-    first_ym = events_sorted[0]["ym"]
-    last_ym = events_sorted[-1]["ym"]
-
-    # Alert level distribution (among event months only)
-    alert_counts: Dict[str, int] = {}
-    for e in events_sorted:
-        if e["occurred"] and e["alertlevel"]:
-            level = e["alertlevel"]
-            alert_counts[level] = alert_counts.get(level, 0) + 1
-
-    # Seasonal frequency: group by calendar month
-    by_cal_month: Dict[int, Dict[str, int]] = {
-        m: {"total": 0, "events": 0} for m in range(1, 13)
-    }
-    for e in events_sorted:
-        parsed = _parse_month_key(e["ym"])
-        if parsed:
-            by_cal_month[parsed.month]["total"] += 1
-            if e["occurred"]:
-                by_cal_month[parsed.month]["events"] += 1
-
-    seasonal: Dict[int, Dict[str, Any]] = {}
-    for cal_month in range(1, 13):
-        total = by_cal_month[cal_month]["total"]
-        evts = by_cal_month[cal_month]["events"]
-        seasonal[cal_month] = {
-            "years_observed": total,
-            "years_with_event": evts,
-            "frequency_pct": round(evts / total * 100, 0) if total > 0 else 0.0,
-        }
-
-    # Recent 12 months
-    recent_12 = events_sorted[-12:] if len(events_sorted) >= 12 else events_sorted[:]
-
-    return {
-        "type": "gdacs_event_history",
-        "iso3": iso3_up,
-        "hazard_code": hz_up,
-        "data_range": f"{first_ym} to {last_ym}",
-        "total_months": total_months,
-        "event_months": event_months,
-        "event_rate_pct": round(event_rate * 100, 1),
-        "alert_distribution": alert_counts,
-        "seasonal": seasonal,
-        "recent_12": recent_12,
-    }
-
-
-def _format_base_rate_for_prompt(
-    history_summary: Dict[str, Any],
-    forecast_keys: list[str],
-    iso3: str = "",
-    hazard_code: str = "",
-) -> str:
-    """Format a base-rate dict into a human-readable prompt block.
-
-    Handles seasonal_profile, conflict_trajectory, no_base_rate, and
-    legacy dict formats (falls back to JSON dump).
-    """
-    summary_type = history_summary.get("type", "")
-    iso3_up = (iso3 or "").upper().strip()
-    hz_up = (hazard_code or "").upper().strip()
-
-    country_name = _load_country_names().get(iso3_up, iso3_up)
-    hazard_display = _HAZARD_DISPLAY_NAMES.get(hz_up, hz_up)
-
-    # Determine which calendar months the forecast covers
-    forecast_months: list[int] = []
-    for k in forecast_keys:
-        parsed = _parse_month_key(k)
-        if parsed:
-            forecast_months.append(parsed.month)
-
-    _MONTH_NAMES = {
-        1: "January", 2: "February", 3: "March", 4: "April",
-        5: "May", 6: "June", 7: "July", 8: "August",
-        9: "September", 10: "October", 11: "November", 12: "December",
-    }
-
-    def _fmt(n: Any) -> str:
-        """Format a number with comma separators, or return 'N/A'."""
-        if n is None:
-            return "N/A"
-        try:
-            return f"{int(n):,}"
-        except (TypeError, ValueError):
-            return str(n)
-
-    if summary_type == "fewsnet_phase3":
-        source = history_summary.get("source", "FEWSNET_IPC")
-        observed = history_summary.get("observed_months", 0)
-        total = history_summary.get("history_length_months", 36)
-        coverage = history_summary.get("coverage_pct", 0)
-        recent_mean = history_summary.get("recent_mean")
-        recent_max = history_summary.get("recent_max")
-        trend = history_summary.get("trend", "unknown")
-        trend_pct = history_summary.get("trend_pct")
-        data_quality = history_summary.get("data_quality", "unknown")
-        last_6m = history_summary.get("last_6m_values", [])
-
-        lines = [
-            "RESOLVER HISTORY (FEWS NET IPC Phase 3+, Current Situation):",
-            f"Phase 3+ population reported in {observed} of the last {total} months ({coverage:.0f}% coverage).",
-        ]
-
-        # Latest value
-        latest = None
-        for entry in reversed(last_6m):
-            if entry.get("value") is not None:
-                latest = entry
-                break
-        if latest:
-            lines.append(f"Latest value: {latest['ym']}: {_fmt(latest['value'])}")
-
-        if recent_mean is not None:
-            lines.append(f"Recent 6-month average (observed only): {_fmt(recent_mean)}")
-        if recent_max is not None:
-            lines.append(f"Peak (recent 6 months): {_fmt(recent_max)}")
-
-        trend_str = trend
-        if trend_pct is not None:
-            trend_str = f"{trend} ({'+' if trend_pct > 0 else ''}{trend_pct:.0f}% over 12 months)"
-        lines.append(f"Trend: {trend_str}")
-        lines.append(f"Data quality: {data_quality} ({coverage:.0f}% monthly coverage)")
-
-        lines.append("")
-        lines.append("Recent values (null = no FEWS NET assessment that month):")
-
-        # Format last 6 months in 2 rows of 3
-        for i in range(0, len(last_6m), 3):
-            chunk = last_6m[i:i+3]
-            parts = []
-            for entry in chunk:
-                val = "null" if entry.get("value") is None else _fmt(entry["value"])
-                parts.append(f"  {entry['ym']}: {val:<12}")
-            lines.append(" | ".join(parts))
-
-        lines.append("")
-        lines.append(
-            "Note: Months marked 'null' mean FEWS NET did not publish a Current "
-            "Situation assessment for this country that month. This does NOT mean "
-            "zero food insecurity. Your forecast should interpolate through gaps "
-            "using the surrounding observed values and seasonal patterns."
-        )
-
-        return "\n".join(lines)
-
-    if summary_type == "seasonal_profile":
-        source = history_summary.get("source", "IFRC")
-        data_range = history_summary.get("data_range", "unknown")
-        years = history_summary.get("years_of_data", 0)
-        months_data = history_summary.get("months", {})
-
-        lines = [
-            f"BASE RATE: {country_name} {hazard_display} ({hz_up}) people affected "
-            f"— {source} data from {data_range} ({years} years)",
-            "",
-            "Seasonal profile for your forecast months:",
-        ]
-
-        # Find the longest month name for alignment
-        month_names_for_forecast = [_MONTH_NAMES.get(m, str(m)) for m in forecast_months]
-        max_name_len = max((len(n) for n in month_names_for_forecast), default=10)
-
-        for cal_month in forecast_months:
-            month_name = _MONTH_NAMES.get(cal_month, str(cal_month))
-            m_data = months_data.get(cal_month) or months_data.get(str(cal_month)) or {}
-            n_obs = m_data.get("n_observations", 0)
-            lines.append(
-                f"  {month_name + ':':<{max_name_len + 1}} "
-                f"min={_fmt(m_data.get('min', 0))}, "
-                f"max={_fmt(m_data.get('max', 0))}, "
-                f"avg={_fmt(m_data.get('mean', 0))}, "
-                f"median={_fmt(m_data.get('median', 0))} "
-                f"({n_obs} obs)"
-            )
-
-        lines.append("")
-        lines.append(
-            "Note: These are historical monthly values across all available years. "
-            "Use as your prior anchor, then update with current signals."
-        )
-        return "\n".join(lines)
-
-    if summary_type == "conflict_trajectory":
-        fat = history_summary.get("fatalities", {})
-        disp = history_summary.get("displacements", {})
-
-        lines = [
-            f"BASE RATE: {country_name} {hazard_display} ({hz_up}) — recent trajectory",
-        ]
-
-        # Fatalities block
-        lines.append("")
-        if fat.get("last_month") is not None:
-            last_f = fat["last_month"]
-            trend_pct = fat.get("trend_pct")
-            trend_dir = fat.get("trend_direction", "unknown")
-
-            if trend_pct == "new_activity":
-                trend_str = "new activity (no prior baseline)"
-            elif trend_pct is not None:
-                trend_str = f"{trend_dir} ({'+' if trend_pct > 0 else ''}{trend_pct}% vs prior 3-month window)"
-            else:
-                trend_str = "insufficient data for trend"
-
-            lines.append(f"Fatalities ({fat.get('source', 'ACLED')}):")
-            lines.append(f"  Last month ({last_f.get('ym', '?')}): {_fmt(last_f.get('value'))}")
-            lines.append(f"  3-month avg: {_fmt(fat.get('trailing_3m_avg'))}/month")
-            lines.append(f"  Trend: {trend_str}")
-        else:
-            note = fat.get("note", "No ACLED fatalities data available.")
-            lines.append(f"Fatalities (ACLED): {note}")
-
-        # Displacements block
-        lines.append("")
-        if disp.get("last_month") is not None:
-            last_d = disp["last_month"]
-            trend_pct = disp.get("trend_pct")
-            trend_dir = disp.get("trend_direction", "unknown")
-
-            if trend_pct == "new_activity":
-                trend_str = "new activity (no prior baseline)"
-            elif trend_pct is not None:
-                trend_str = f"{trend_dir} ({'+' if trend_pct > 0 else ''}{trend_pct}% vs prior 3-month window)"
-            else:
-                trend_str = "insufficient data for trend"
-
-            lines.append(f"Displacement ({disp.get('source', 'IDMC')}):")
-            lines.append(f"  Last month ({last_d.get('ym', '?')}): {_fmt(last_d.get('value'))} new displacements")
-            lines.append(f"  3-month avg: {_fmt(disp.get('trailing_3m_avg'))}/month")
-            lines.append(f"  Trend: {trend_str}")
-        else:
-            note = disp.get("note", "No IDMC displacement data available.")
-            lines.append(f"Displacement (IDMC): {note}")
-
-        lines.append("")
-        lines.append(
-            "Note: Conflict base rates reflect recent trajectory, not seasonal patterns. "
-            "Use as your prior anchor."
-        )
-        return "\n".join(lines)
-
-    if summary_type == "no_base_rate":
-        note = history_summary.get("note", "No base rate available for this hazard.")
-        return (
-            f"BASE RATE: {country_name} {hazard_display} ({hz_up})\n"
-            f"{note}"
-        )
-
-    # Legacy fallback: dump as JSON (backward compat for any unrecognised type)
-    return (
-        "Resolver history summary (Resolver is one imperfect source; "
-        "ACLED strong, IDMC short, IFRC Montandon may be sparse):\n"
-        "```json\n"
-        f"{json.dumps(history_summary, default=str, indent=2)}\n"
-        "```"
-    )
 
 
 def _load_fewsnet_phase3_history(
@@ -1339,83 +867,6 @@ def _must_dict(name: str, obj: Any) -> Dict[str, Any]:
     return d
 
 
-def _coerce_date(val: Any) -> Optional[date]:
-    if isinstance(val, date):
-        return val
-    if isinstance(val, str):
-        try:
-            return datetime.fromisoformat(val).date()
-        except Exception:
-            return None
-    return None
-
-
-def _parse_month_key(s: str) -> Optional[date]:
-    """
-    Parse a month key into a date anchored to the first of the month.
-
-    Supported formats:
-      - YYYY-MM
-      - YYYY-MM-DD (day component is discarded)
-    """
-    if not isinstance(s, str):
-        return None
-
-    text = s.strip()
-    if not text:
-        return None
-
-    try:
-        if len(text) == 7:
-            dt = datetime.strptime(text, "%Y-%m")
-            return date(dt.year, dt.month, 1)
-        dt = datetime.fromisoformat(text)
-        return date(dt.year, dt.month, 1)
-    except Exception:
-        try:
-            parts = text.split("-")
-            if len(parts) >= 2:
-                year = int(parts[0])
-                month = int(parts[1])
-                return date(year, month, 1)
-        except Exception:
-            return None
-    return None
-
-
-def _sanitize_month_series(
-    month_to_value: Dict[str, Any],
-) -> tuple[Dict[str, Any], list[str], list[str]]:
-    """
-    Remove entries with month keys later than the current month.
-
-    Returns (cleaned_dict, dropped_future_months, unparseable_month_keys).
-    """
-    now_month = datetime.now(timezone.utc).replace(tzinfo=None).strftime("%Y-%m")
-    cleaned: Dict[str, Any] = {}
-    dropped: list[str] = []
-    unparseable: list[str] = []
-
-    for key in sorted(month_to_value.keys()):
-        val = month_to_value[key]
-        if str(key).strip() == "":
-            unparseable.append(str(key))
-            continue
-        parsed = _parse_month_key(str(key))
-        if parsed is None:
-            unparseable.append(str(key))
-            continue
-
-        month_key = parsed.strftime("%Y-%m")
-        if month_key > now_month:
-            dropped.append(str(key))
-            continue
-
-        cleaned[str(key)] = val
-
-    return cleaned, dropped, unparseable
-
-
 def _build_month_labels(start_date: Optional[date], horizon_months: int = 6) -> Dict[int, str]:
     labels: Dict[int, str] = {}
     if not isinstance(start_date, date):
@@ -1624,55 +1075,6 @@ def _sanitize_markdown_chunks(chunks: List[Any]) -> List[str]:
             # happened, and failing to write the human log is worse.
             continue
     return sanitized
-
-
-def _pythia_db_url_from_config() -> str:
-    """
-    Best-effort helper to read the Pythia DuckDB URL from config or env.
-
-    Priority:
-      1. pythia.db.schema.get_db_url (if available)
-      2. app.db_url from pythia.config
-      3. PYTHIA_DB_URL environment variable
-      4. default duckdb:///data/resolver.duckdb
-
-    This helper is intentionally kept for backward compatibility with tests
-    that monkeypatch it to point to a temporary DuckDB file.
-    """
-
-    try:
-        from pythia.db.schema import get_db_url
-
-        url = get_db_url()
-        if url:
-            return url
-    except Exception:
-        pass
-
-    if _PYTHIA_CFG_LOAD is not None:
-        try:
-            cfg = _PYTHIA_CFG_LOAD()
-            app_cfg = cfg.get("app", {}) if isinstance(cfg, dict) else {}
-            db_url = str(app_cfg.get("db_url", "")).strip()
-            if db_url:
-                return db_url
-        except Exception:
-            pass
-
-    env_url = os.getenv("PYTHIA_DB_URL", "").strip()
-    if env_url:
-        return env_url
-
-    return "duckdb:///data/resolver.duckdb"
-
-
-def _pythia_db_path_from_config() -> str:
-    """Return a filesystem path for the configured DuckDB database."""
-
-    db_url = _pythia_db_url_from_config()
-    if db_url.startswith("duckdb:///"):
-        return db_url.replace("duckdb:///", "", 1)
-    return db_url
 
 
 def _select_hs_run_id_for_forecast(
@@ -2075,358 +1477,6 @@ def _write_spd_members_v2_to_db(
             con.close()
         except Exception:
             pass
-
-def _load_ifrc_pa_history(
-    iso3: str,
-    hazard_code: str,
-    *,
-    months: int = 36,
-) -> Tuple[str, Dict[str, Any]]:
-    """
-    Load a 36-month IFRC Montandon 'people affected' history for a given
-    ISO3 + Pythia hazard code.
-
-    Queries facts_resolved (where IFRC Montandon connector data lands)
-    for natural hazard PA metrics.
-    """
-
-    hz = (hazard_code or "").upper()
-
-    try:
-        con = duckdb.connect(_pythia_db_path_from_config(), read_only=True)
-    except Exception:
-        return "", {"error": "missing_db", "history_rows_detail": [], "summary_text": ""}
-
-    try:
-        rows = con.execute(
-            """
-            SELECT ym, value, COALESCE(source_id, '') AS source_id
-            FROM facts_resolved
-            WHERE iso3 = ?
-              AND hazard_code = ?
-              AND lower(metric) IN ('affected', 'in_need', 'pa')
-            ORDER BY ym DESC
-            LIMIT ?
-            """,
-            [iso3, hz, months],
-        ).fetchall()
-    except Exception as exc:
-        con.close()
-        return "", {
-            "error": f"ifrc_query_error:{type(exc).__name__}",
-            "history_rows_detail": [],
-            "summary_text": "",
-        }
-
-    con.close()
-
-    if not rows:
-        return "", {"error": "no_history", "history_rows_detail": [], "summary_text": ""}
-
-    history: List[Dict[str, Any]] = []
-    values: List[float] = []
-    for ym, pa_val, source_id in rows:
-        ym_str = str(ym)
-        v = float(pa_val or 0.0)
-        history.append(
-            {
-                "ym": ym_str,
-                "value": v,
-                "source": "IFRC",
-                "source_id": source_id,
-            }
-        )
-        values.append(v)
-
-    history_for_table = list(reversed(history))
-
-    summary_lines: List[str] = [
-        "### IFRC Montandon people affected — 36-month history (Resolver)",
-        "",
-        f"- Months available: {len(history)}",
-        f"- Min monthly PA: {min(values):,.0f}",
-        f"- Max monthly PA: {max(values):,.0f}",
-    ]
-    summary_lines.append("")
-    summary_lines.append("| Month | People affected |")
-    summary_lines.append("|-------|-----------------|")
-    for row in history_for_table:
-        summary_lines.append(f"| {row['ym']} | {row['value']:,.0f} |")
-    summary = "\n".join(summary_lines)
-
-    return summary, {
-        "error": "",
-        "history_rows_detail": history,
-        "summary_text": summary,
-    }
-
-
-def _load_idmc_pa_history(
-    iso3: str,
-    hazard_code: str,
-    *,
-    months: int = 36,
-) -> Tuple[str, Dict[str, Any]]:
-    """
-    Load a 36-month IDMC/DTM displacement history for conflict/displacement
-    questions.
-
-    Combines:
-      - flows from facts_deltas (new_displacements, idp_displacement_new_dtm),
-      - stocks from facts_resolved (idp_displacement_stock_dtm).
-
-    We group by ym and sum flows; stocks are carried as-is per ym.
-    """
-
-    hz = (hazard_code or "").upper()
-    if hz not in IDMC_HZ_MAP:
-        return "", {"error": "no_mapping", "history_rows_detail": [], "summary_text": ""}
-
-    try:
-        con = duckdb.connect(_pythia_db_path_from_config(), read_only=True)
-    except Exception:
-        return "", {"error": "missing_db", "history_rows_detail": [], "summary_text": ""}
-
-    try:
-        flow_rows = con.execute(
-            """
-            SELECT
-                ym,
-                SUM(CASE WHEN lower(metric) = 'new_displacements'
-                         THEN COALESCE(value_new, 0) ELSE 0 END) AS idmc_flow,
-                SUM(CASE WHEN lower(metric) = 'idp_displacement_new_dtm'
-                         THEN COALESCE(value_new, 0) ELSE 0 END) AS dtm_flow
-            FROM facts_deltas
-            WHERE iso3 = ?
-              AND lower(series_semantics) = 'new'
-              AND lower(metric) IN ('new_displacements', 'idp_displacement_new_dtm')
-            GROUP BY ym
-            ORDER BY ym DESC
-            LIMIT ?
-            """,
-            [iso3, months],
-        ).fetchall()
-
-        stock_rows = con.execute(
-            """
-            SELECT ym, value AS stock_value
-            FROM facts_resolved
-            WHERE iso3 = ?
-              AND lower(metric) = 'idp_displacement_stock_dtm'
-            ORDER BY ym DESC
-            LIMIT ?
-            """,
-            [iso3, months],
-        ).fetchall()
-    except Exception as exc:
-        con.close()
-        return "", {
-            "error": f"idmc_query_error:{type(exc).__name__}",
-            "history_rows_detail": [],
-            "summary_text": "",
-        }
-
-    if not flow_rows and not stock_rows:
-        con.close()
-        return "", {"error": "no_history", "history_rows_detail": [], "summary_text": ""}
-
-    stock_by_ym = {str(ym): float(stock or 0.0) for ym, stock in stock_rows}
-
-    history: List[Dict[str, Any]] = []
-    chosen_flow_values: List[float] = []
-    stock_values: List[float] = []
-
-    for ym, idmc_flow, dtm_flow in flow_rows:
-        ym_str = str(ym)
-        idmc_v = float(idmc_flow or 0.0)
-        dtm_v = float(dtm_flow or 0.0)
-        if idmc_v:
-            chosen = idmc_v
-            source = "IDMC"
-        elif dtm_v:
-            chosen = dtm_v
-            source = "DTM"
-        else:
-            chosen = 0.0
-            source = ""
-
-        stock_v = stock_by_ym.get(ym_str, 0.0)
-
-        history.append(
-            {
-                "ym": ym_str,
-                "value": chosen,
-                "value_flow": chosen,
-                "value_stock": stock_v,
-                "source": source or "IDMC/DTM",
-                "flow_idmc": idmc_v,
-                "flow_dtm": dtm_v,
-            }
-        )
-        chosen_flow_values.append(chosen)
-        stock_values.append(stock_v)
-
-    con.close()
-
-    if not history and not stock_values:
-        return "", {"error": "no_history", "history_rows_detail": [], "summary_text": ""}
-
-    history_for_table = list(reversed(history))
-
-    summary_lines: List[str] = [
-        "### IDMC/DTM displacement — 36-month history (Resolver)",
-        "",
-        f"- Months available: {len(history)}",
-    ]
-    if chosen_flow_values:
-        summary_lines.append(
-            f"- Flow (new displacements) min={min(chosen_flow_values):,.0f}, max={max(chosen_flow_values):,.0f}"
-        )
-    if stock_values:
-        summary_lines.append(
-            f"- Stock (IDPs, DTM) min={min(stock_values):,.0f}, max={max(stock_values):,.0f}"
-        )
-    summary_lines.append(
-        "- Source selection: IDMC is used when available for a given month; DTM is used only when IDMC is missing."
-    )
-    summary_lines.append("")
-    summary_lines.append("| Month | Flow (people displaced) | Stock (IDPs, DTM) | Source |")
-    summary_lines.append("|-------|--------------------------|-------------------|--------|")
-    for row in history_for_table:
-        summary_lines.append(
-            f"| {row['ym']} | {row['value_flow']:,.0f} | {row['value_stock']:,.0f} | {row['source']} |"
-        )
-    summary = "\n".join(summary_lines)
-
-    return summary, {
-        "error": "",
-        "history_rows_detail": history,
-        "summary_text": summary,
-    }
-
-
-def _load_acled_fatalities_history(
-    iso3: str,
-    *,
-    months: int = 36,
-) -> Tuple[str, Dict[str, Any]]:
-    """
-    Load a 36-month ACLED fatalities history for conflict questions.
-
-    First try facts_resolved with metric='fatalities'; fall back to
-    db.acled_monthly_fatalities if needed.
-    """
-
-    try:
-        con = duckdb.connect(_pythia_db_path_from_config(), read_only=True)
-    except Exception:
-        return "", {"error": "missing_db", "history_rows_detail": [], "summary_text": ""}
-
-    try:
-        rows = con.execute(
-            """
-            SELECT ym, value, source_type
-            FROM facts_resolved
-            WHERE iso3 = ?
-              AND lower(metric) = 'fatalities'
-            ORDER BY ym DESC
-            LIMIT ?
-            """,
-            [iso3, months],
-        ).fetchall()
-    except Exception as exc:
-        rows = []
-        facts_err = f"facts_query_error:{type(exc).__name__}"
-    else:
-        facts_err = ""
-
-    history: List[Dict[str, Any]] = []
-    values: List[float] = []
-
-    if rows:
-        for ym, val, source_type in rows:
-            ym_str = str(ym)
-            v = float(val or 0)
-            history.append({"ym": ym_str, "value": v, "source": source_type or "acled"})
-            values.append(v)
-    else:
-        try:
-            rows2 = con.execute(
-                """
-                SELECT strftime(month, '%Y-%m') AS ym, fatalities, source
-                FROM acled_monthly_fatalities
-                WHERE iso3 = ?
-                ORDER BY month DESC
-                LIMIT ?
-                """,
-                [iso3, months],
-            ).fetchall()
-        except Exception as exc:
-            con.close()
-            return "", {
-                "error": facts_err or f"acled_query_error:{type(exc).__name__}",
-                "history_rows_detail": [],
-                "summary_text": "",
-            }
-
-        if not rows2:
-            con.close()
-            return "", {"error": "no_rows", "history_rows_detail": [], "summary_text": ""}
-
-        for ym_str, fatalities, src in rows2:
-            v = float(fatalities or 0)
-            history.append({"ym": ym_str, "value": v, "source": src or "acled"})
-            values.append(v)
-
-    con.close()
-
-    values_rev = list(reversed(values))
-    summary_lines = [
-        "### ACLED conflict fatalities — 36-month history (Resolver)",
-        "",
-        f"- Months available: {len(history)}",
-        f"- Min monthly fatalities: {min(values_rev):,.0f}",
-        f"- Max monthly fatalities: {max(values_rev):,.0f}",
-    ]
-    summary = "\n".join(summary_lines)
-
-    return summary, {
-        "error": "",
-        "history_rows_detail": history,
-        "summary_text": summary,
-    }
-
-
-def _load_pa_history_block(
-    iso3: str,
-    hazard_code: str,
-    *,
-    metric: str,
-    months: int = 36,
-) -> tuple[str, Dict[str, Any]]:
-    """
-    Dispatch to the appropriate history loader based on metric + hazard.
-
-    - For metric='FATALITIES' → ACLED fatalities history.
-    - For metric='PA' and conflict/displacement hazards (ACO/ACE/CU/DI) →
-      IDMC/DTM displacement history.
-    - For metric='PA' and natural hazards (FL/DR/TC/HW) → IFRC Montandon PA history.
-    """
-
-    hz = (hazard_code or "").upper()
-    m = (metric or "").upper()
-
-    if m == "FATALITIES":
-        return _load_acled_fatalities_history(iso3, months=months)
-
-    if m == "PA" and hz in IDMC_HZ_MAP:
-        return _load_idmc_pa_history(iso3, hazard_code, months=months)
-
-    if m == "PA" and hz in NATURAL_HAZARD_CODES:
-        return _load_ifrc_pa_history(iso3, hazard_code, months=months)
-
-    return "", {"error": "no_mapping", "history_rows_detail": [], "summary_text": ""}
-
 
 def _safe_row_get(row: Any, index: int, name: str) -> Any:
     """
@@ -3144,22 +2194,6 @@ def _retriever_enabled() -> bool:
 def _retriever_model_id() -> str | None:
     model_id = (os.getenv("PYTHIA_RETRIEVER_MODEL_ID") or "").strip()
     return model_id or None
-
-
-@functools.lru_cache(maxsize=1)
-def _load_country_names() -> dict[str, str]:
-    names: dict[str, str] = {}
-    try:
-        with COUNTRIES_CSV.open("r", encoding="utf-8-sig", newline="") as handle:
-            reader = csv.DictReader(handle)
-            for row in reader:
-                iso3 = (row.get("iso3") or "").strip().upper()
-                name = (row.get("country_name") or "").strip()
-                if iso3 and name:
-                    names[iso3] = name
-    except Exception:
-        return {}
-    return names
 
 
 def _country_label(iso3: str | None) -> str:
@@ -4021,64 +3055,6 @@ async def _call_spd_ensemble_v2(
     return spd_obj, aggregated_usage, raw_calls
 
 
-def _is_calendar_month_key(key: str) -> bool:
-    try:
-        import re as _re  # local import to keep global imports minimal
-
-        return bool(_re.match(r"^\d{4}-\d{2}$", str(key).strip()))
-    except Exception:
-        return False
-
-
-def _parse_month_offset_key(key: str) -> int | None:
-    """
-    Parse month offset keys like 'month_0', 'month_1', 'm0', 'm1'.
-    Returns the integer offset or None if not recognized.
-    """
-    if not isinstance(key, str):
-        return None
-    k = key.strip().lower()
-    if k.startswith("month_") and k[6:].isdigit():
-        try:
-            return int(k[6:])
-        except Exception:
-            return None
-    if k.startswith("m") and k[1:].isdigit():
-        try:
-            return int(k[1:])
-        except Exception:
-            return None
-    return None
-
-
-def _add_months(ym: str, offset: int) -> str:
-    """Return YYYY-MM shifted by offset months (offset can be negative)."""
-    parts = str(ym or "").split("-")
-    if len(parts) != 2:
-        return ""
-    try:
-        y = int(parts[0])
-        m = int(parts[1])
-    except Exception:
-        return ""
-    total_months = (y * 12 + (m - 1)) + int(offset)
-    year = total_months // 12
-    month = (total_months % 12) + 1
-    return f"{year:04d}-{month:02d}"
-
-
-def _expected_months(anchor_month: str, n: int = NUM_HORIZONS) -> list[str]:
-    """Return the n forecast-window months starting at ``anchor_month``.
-
-    ``anchor_month`` is the FIRST window month (horizon_m=1) — see
-    ``_anchor_month_for_question``. It must never be the questions-table
-    ``target_month`` (the 6th window month).
-    """
-    if not anchor_month:
-        return []
-    return [_add_months(anchor_month, i) for i in range(n)]
-
-
 def _build_bayesmc_spd_obj(
     per_model_spds: list[dict[str, list[float]]],
     *,
@@ -4825,79 +3801,6 @@ def _normalize_question_row_for_spd(question_row: Any) -> Dict[str, Any]:
     raise TypeError(
         f"Unsupported question_row type {type(question_row)!r}; expected Mapping, duckdb.Row, or tuple/list."
     )
-
-
-def _first_target_month(target_months: Any) -> str | None:
-    """Return the first target month string from a string or iterable, if present."""
-
-    if isinstance(target_months, str) and target_months.strip():
-        return target_months.strip()
-
-    if isinstance(target_months, (list, tuple)):
-        for month_val in target_months:
-            if isinstance(month_val, str) and month_val.strip():
-                return month_val.strip()
-
-    return None
-
-
-def _anchor_month_for_question(rec: Mapping[str, Any]) -> str | None:
-    """Return the first forecast-window month ('YYYY-MM') for a question row.
-
-    ``window_start_date`` is authoritative: compute_resolutions maps
-    horizon_m=1 to the window_start month, so month labels in prompts and
-    month-offset expansion must anchor there. ``target_month`` in the
-    questions table is the 6th (last) window month, so when window_start
-    is missing the anchor is target_month minus 5 months. Anchoring at
-    target_month directly shifts every forecast +5 months (the bug that
-    affected runs 2026-03-21 → 2026-07-01).
-    """
-    ws = _coerce_date(rec.get("window_start_date"))
-    if ws is not None:
-        return f"{ws.year:04d}-{ws.month:02d}"
-    tm = _first_target_month(rec.get("target_months") or rec.get("target_month"))
-    if tm:
-        anchored = _add_months(tm[:7], -5)
-        return anchored or None
-    return None
-
-
-def _month_index_for_label(label: str, anchor_month: str | None) -> int | None:
-    """Map a forecast month label to its 1-based horizon index (1..6).
-
-    Calendar labels ('YYYY-MM' / 'YYYY-MM-DD') are offset from
-    ``anchor_month`` (the first window month, = resolutions horizon_m=1);
-    canonical 'month_N' labels map to N directly. Returns None for labels
-    that fall outside the 6-month window or cannot be parsed — positional
-    enumeration must never be used instead, because a missing or off-window
-    label would silently shift every subsequent month against resolutions.
-    """
-    s = str(label).strip()
-
-    def _cal_index(y: int, m: int) -> int | None:
-        if not anchor_month:
-            return None
-        try:
-            ay, am = map(int, anchor_month.split("-"))
-        except Exception:
-            return None
-        idx = (y * 12 + m) - (ay * 12 + am) + 1
-        return idx if 1 <= idx <= NUM_HORIZONS else None
-
-    if _is_calendar_month_key(s):
-        y, m = map(int, s.split("-"))
-        return _cal_index(y, m)
-
-    offset = _parse_month_offset_key(s)
-    if offset is not None:
-        idx = offset if offset >= 1 else 1  # 'month_0' style → first month
-        return idx if 1 <= idx <= NUM_HORIZONS else None
-
-    dt = _parse_month_key(s)
-    if dt is not None:
-        return _cal_index(dt.year, dt.month)
-
-    return None
 
 
 # ---------------------------------------------------------------------------
@@ -6273,119 +5176,6 @@ def _maybe_dump_raw_gtmc1(content: str, *, run_id: str, question_id: str) -> Opt
         return path
     except Exception:
         return None
-
-# --------------------------------------------------------------------------------
-# Calibration weights loader (optional legacy file fallback).
-# --------------------------------------------------------------------------------
-
-def _load_calibration_weights_file() -> Dict[str, Any]:
-    path = os.getenv("CALIB_WEIGHTS_PATH", "")
-    if not path or not os.path.exists(path):
-        return {}
-    try:
-        with open(path, "r", encoding="utf-8") as f:
-            return json.load(f)
-    except Exception:
-        return {}
-
-
-def _load_calibration_weights_db(
-    hazard_code: str,
-    metric: str,
-) -> Optional[Dict[str, float]]:
-    try:
-        from resolver.db import duckdb_io
-    except Exception:
-        return None
-
-    hz = (hazard_code or "").upper().strip()
-    mt = (metric or "").upper().strip()
-    if not hz or not mt:
-        return None
-
-    db_url = _pythia_db_url_from_config() or os.getenv("RESOLVER_DB_URL", "").strip()
-    if not db_url:
-        return None
-
-    conn = None
-    try:
-        conn = duckdb_io.get_db(db_url)
-    except Exception:
-        return None
-
-    try:
-        row = conn.execute(
-            """
-            SELECT as_of_month
-            FROM calibration_weights
-            WHERE hazard_code = ? AND metric = ?
-            ORDER BY as_of_month DESC
-            LIMIT 1
-            """,
-            [hz, mt],
-        ).fetchone()
-        if not row:
-            return None
-        as_of_month = str(row[0])
-
-        rows = conn.execute(
-            """
-            SELECT model_name, weight
-            FROM calibration_weights
-            WHERE hazard_code = ? AND metric = ? AND as_of_month = ?
-            ORDER BY COALESCE(model_name, '')
-            """,
-            [hz, mt, as_of_month],
-        ).fetchall()
-        if not rows:
-            return None
-
-        weights: Dict[str, float] = {}
-        for model_name, weight in rows:
-            if model_name is None:
-                continue
-            weights[str(model_name)] = float(weight)
-        if not weights:
-            return None
-
-        if os.getenv("PYTHIA_DEBUG_DB", "0") == "1":
-            print(
-                "[forecaster] loaded calibration weights for hazard="
-                f"{hz} metric={mt} as_of={as_of_month}: {weights}"
-            )
-        return weights
-    except Exception:
-        return None
-    finally:
-        try:
-            duckdb_io.close_db(conn)
-        except Exception:
-            pass
-
-
-def _calibration_weights_enabled() -> bool:
-    """Whether calibration weights are applied to ensemble aggregation."""
-    return os.getenv("PYTHIA_USE_CALIBRATION_WEIGHTS", "1") != "0"
-
-
-# Per-process cache: one DB read per (hazard, metric) per run instead of one
-# per question. Values may be None (no weights available). Guarded by a lock:
-# the SPD phase runs questions on worker threads, and an unguarded check-then-
-# set raced duplicate DB reads for the same (hazard, metric).
-_CALIB_WEIGHTS_CACHE: Dict[Tuple[str, str], Optional[Dict[str, float]]] = {}
-_CALIB_WEIGHTS_LOCK = threading.Lock()
-
-
-def _reset_calibration_weights_cache() -> None:
-    """Clear cached calibration weights (call at run start so a long-lived
-    process picks up freshly computed weights)."""
-    with _CALIB_WEIGHTS_LOCK:
-        _CALIB_WEIGHTS_CACHE.clear()
-
-
-def _member_weight_key(ms: ModelSpec) -> str:
-    """Disambiguated member key matching _write_spd_members_v2_to_db naming."""
-    return f"{getattr(ms, 'name', '')} ({getattr(ms, 'model_id', '')})"
 
 
 def _resolve_member_weights(
